requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.39.0
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.28.1
